Fix code quality findings.
