+++ Core/Core/Resources/Database.xcdatamodeld/Database.xcdatamodel/contents
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="23788" systemVersion="24D70" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="23788" systemVersion="24E248" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="AccountNotification" representedClassName="Core.AccountNotification" syncable="YES">
         <attribute name="endAt" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <attribute name="iconRaw" attributeType="String"/>
@@ -210,20 +210,20 @@
     <entity name="CDRubricCriterion" representedClassName="Core.CDRubricCriterion" syncable="YES">
         <attribute name="assignmentID" attributeType="String"/>
         <attribute name="criterionUseRange" attributeType="Boolean" defaultValueString="NO" usesScalarValueType="YES"/>
-        <attribute name="desc" attributeType="String"/>
         <attribute name="id" attributeType="String"/>
         <attribute name="ignoreForScoring" attributeType="Boolean" defaultValueString="NO" usesScalarValueType="YES"/>
-        <attribute name="longDesc" attributeType="String"/>
+        <attribute name="longDescription" attributeType="String"/>
         <attribute name="points" attributeType="Double" defaultValueString="0.0" usesScalarValueType="YES"/>
+        <attribute name="shortDescription" attributeType="String"/>
         <relationship name="assignment" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="Assignment" inverseName="rubricRaw" inverseEntity="Assignment"/>
         <relationship name="ratingsRaw" optional="YES" toMany="YES" deletionRule="Nullify" ordered="YES" destinationEntity="CDRubricRating" inverseName="rubric" inverseEntity="CDRubricRating"/>
     </entity>
     <entity name="CDRubricRating" representedClassName="Core.CDRubricRating" syncable="YES">
         <attribute name="assignmentID" attributeType="String"/>
-        <attribute name="desc" attributeType="String"/>
         <attribute name="id" attributeType="String"/>
-        <attribute name="longDesc" attributeType="String"/>
+        <attribute name="longDescription" attributeType="String"/>
         <attribute name="points" attributeType="Double" defaultValueString="0.0" usesScalarValueType="YES"/>
+        <attribute name="shortDescription" attributeType="String"/>
         <relationship name="rubric" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="CDRubricCriterion" inverseName="ratingsRaw" inverseEntity="CDRubricCriterion"/>
     </entity>
     <entity name="CDWidgetAnnouncement" representedClassName="Core.CDWidgetAnnouncement" syncable="YES">
